requirements.txt
torch>=2.0.0
torchvision>=0.15.0
timm>=0.9.0
transformers>=4.35.0
facenet-pytorch>=2.5.3
opencv-python>=4.8.0
numpy>=1.24.0
scikit-learn>=1.3.0
streamlit>=1.28.0
matplotlib>=3.7.0
tqdm>=4.65.0
Pillow>=10.0.0
huggingface_hub>=0.19.0
